% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_epidemics.R
\name{generate_epidemics}
\alias{generate_epidemics}
\title{Generate epidemics using the Gillespie algorithm}
\arguments{
\item{offspring_data}{dataframe created by generate_population function}

\item{group_size}{group size}
}
\value{
return offspring_data dataframe with column containing time of

infection
}
\description{
Generate epidemics using the Gillespie algorithm

assumptions:

* animals with genetic structure and allocated in groups

* SI model

* all individuals are infected during the epidemic (prevalence = 1)
}
\examples{
grp_size = 10
df <- generate_population(group_size = grp_size, sires = 25, dpsire = 2)$offspring
df_epi <- generate_epidemics(df, group_size = grp_size)

}
