% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_population.R
\name{generate_population}
\alias{generate_population}
\title{Generate population of animals with genetic structure and allocated
in groups}
\arguments{
\item{num_replications}{number of replications}

\item{sires}{number of sires}

\item{dpsire}{number of dam per sires}

\item{group_size}{group size}

\item{index}{1 if animal is index case, 0 otherwise}

\item{SigG.g}{= susceptibility genetic variance}

\item{SigG.f}{= infectivity genetic variance}

\item{SigE.g}{= susceptibility environmental variance}

\item{SigE.f}{= infectivity environmental variance}

\item{rhoG}{genetic correlation between susceptibility and infectivity}

\item{rhoE}{environmental correlation between susceptibility and infectivity}
}
\value{
data frame with
}
\description{
Generate population of animals with genetic structure and allocated
in groups
}
\examples{
grp_size = 10
example <- generate_population(num_replications = 10, group_size = grp_size, sires = 25, dpsire = 2,
            rhoG = -0.4, rhoE = -0.4, SigG.g = 4, SigG.f = 4, SigE.g = 1,
             SigE.f = 1)
example


}
